requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
openai==1.12.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.13
